Read remote default branch, fix set-upstream bug.
I can't recall now why it was this was a bug, since I made the change locally and forgot all about it. But at one point there were occaisionally errors with the way I was setting the upstream branch. To address that, I'm just executing the command a bit differently, in a way that pushes the branch and sets the upstream at the same time.

Something that's become an issue more and more recently, are those repositories which don't use the `master` branch as the default branch. In many cases; that branch is named `main` instead. It's entirely possible, the default branch isn't even named that. In some cases, depending on the DLC established, `develop` may very well be the default branch. In any of these cases, the module currently fails, because when we perform operations that require us to jump to the `master` (default) branch there is no such branch to switch to. To combat this, I removed the places where `master` was hard coded into the module, and instead read the remote to determine what the default branch actually is supposed to be.

This also adds a new function to the module, `git-Edit-Commit-Document` (alias `gec`). This command opens a vim editor to the `commit-message.md` file. This allows me to incrementally build out the commit message as I do work. When `git-Stage` (`gst`) is called, and directed to commit; if the `commit-message.md` file exists, then the commit command will be executed to open the editor using the `commit-message.md` as the default contents for editing. This way I can incrementally craft a commit message long before I'm ready to commit, then just review and accept that message on commit. The advantage is that I don't have to take time to build a rich history of the work done in that commit at the very end.